requirements-dev.txt
Lin-CMS==0.4.8
pillow==9.0.1
flask-cors==3.0.10
gunicorn==20.1.0
gevent==21.8.0
flask-socketio==5.1.1
blinker==1.4
python-dotenv==0.19.2
gevent-websocket==0.10.1
flask-sqlacodegen==1.1.8
black==21.12b0
isort==5.10.1
watchdog==2.1.6
coverage==6.3.2
pytest==6.2.5
pre-commit==2.17.0
pytest-ordering==0.6
flake8==4.0.1
xlwt==1.3.0
